Refactor custom explorer templates
Split custom explorer template parsing and validation into a

a dedicated module so public option logic stays focused. Keep

template behavior intact while centralizing placeholder

normalization, URL matching, and stored-template fallback
